requirements.txt
asgiref==3.7.2
blinker==1.7.0
blueprint==3.4.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Login==0.6.3
flask-mailman==1.0.0
Flask-Principal==0.4.0
Flask-Security-Too==5.4.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
icecream==2.1.3
idna==3.6
importlib_resources==6.1.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2==2.9.9
SQLAlchemy==2.0.28
typing_extensions==4.10.0
uvicorn==0.28.0
Werkzeug==3.0.1
WTForms==3.1.2
WTForms-SQLAlchemy==0.4.1
